gemini-chatbot/requirements.txt
google-generativeai 
streamlit==1.24.0
python-dotenv==1.0.1
